backend/requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
Authlib==1.3.0
bcrypt==4.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.120.2
filelock==3.20.0
fsspec==2025.10.0
greenlet==3.2.4
groq==0.33.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.27.0
huggingface-hub==1.0.1
idna==3.11
itsdangerous==2.1.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pillow==11.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.7.1
pydantic_core==2.41.4
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
